requirements.txt
Flask
pandas
numpy
scikit-learn
lightgbm
mlflow
joblib
PyYAML
gunicorn